Fix expected result test for Shuffle.
The explanatory comment incorrectly state that uint64 is 4 bytes, which
made the math very confusing. It also did not work on big-endian
systems, since everything is swapped from what it assumes.

Using an explicit order fixes it on both systems, and big-endian makes
it easier to reason about since the typed-in values look the same as the
layout in memory.